Move service implementation to util class
